Rename
TextFieldWidget -> OneLineTextField
StringTextFieldWidget -> TextFieldWidget
