+++ src/main/java/com/cleanroommc/modularui/test/TestTile.java
@@ -30,7 +30,7 @@ import com.cleanroommc.modularui.widgets.slot.ModularSlot;
 import com.cleanroommc.modularui.widgets.slot.SlotGroup;
 import com.cleanroommc.modularui.widgets.textfield.IntFieldWidget;
 import com.cleanroommc.modularui.widgets.textfield.LongFieldWidget;
-import com.cleanroommc.modularui.widgets.textfield.StringTextFieldWidget;
+import com.cleanroommc.modularui.widgets.textfield.TextFieldWidget;
 
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
@@ -44,7 +44,6 @@ import net.minecraftforge.items.ItemStackHandler;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Function;
 
 public class TestTile extends TileEntity implements IGuiHolder<PosGuiData>, ITickable {
 
@@ -158,7 +157,7 @@ public class TestTile extends TileEntity implements IGuiHolder<PosGuiData>, ITic
                                                         })
                                                         //.flex(flex -> flex.left(3)) // ?
                                                         .overlay(IKey.str("Button 2")))
-                                                .child(new StringTextFieldWidget()
+                                                .child(new TextFieldWidget()
                                                         .size(60, 20)
                                                         .value(SyncHandlers.string(() -> this.value, val -> this.value = val))
                                                         .margin(0, 3))
@@ -373,7 +372,7 @@ public class TestTile extends TileEntity implements IGuiHolder<PosGuiData>, ITic
     public void buildDialog(Dialog<String> dialog) {
         AtomicReference<String> value = new AtomicReference<>("");
         dialog.setDraggable(true);
-        dialog.child(new StringTextFieldWidget()
+        dialog.child(new TextFieldWidget()
                         .flex(flex -> flex.size(100, 20).align(Alignment.Center))
                         .value(new StringValue.Dynamic(value::get, value::set)))
                 .child(new ButtonWidget<>()
